Updated .travis.yml. Added PyPy to the build matrix.
Added pypy and pypy3 to the build.

Install the library so that tests will be able to run in CI.

Set the include path correctly so that libxml2 will be found.

Remove PyPy3 from the matrix for now since it has a very old CFFI.

Moved the python modules into a package since this package no longer contains a single script.

No need to build in CI since the package is already installed.