Fix the majority of PEP8 issues (#41)
* pep8: fix whitespaces in lines

Add or remove whitespaces within lines, adapting the indentation if
needed.  This fixes the following issues:
E124, E201, E202, E211, E221, E231, and E251.

This commit is best viewed with 'git diff -w'.

* pep8: add missing empty lines

This fixes the following issue: E302.

* pep8: fix comparisons with None

Use 'foo is not None' to compare against None.

This fixes the following issues: E711, and E721.

* pep8: wrap lines longer than 80 characters

This fixes the following issue: E501.

* pep8: fix location of imports in Augeas

Move them before other instructions, as recommended.
This fixes the following issue: E402, but not in test_augeas (which
needs to augment sys.path, in case it is run as standalone script).